// crm_frontend_3/src/api/api2.ts
export interface Type_AccessoriesVm {
    type_Сomponent: string;
}

export interface Get_AccessoriesVm {
    сharacteristics_info:string;
     name_Company : string;
}

export interface Accessories_CPUVm {
    iD_Сomponent: string;
    name_Сomponent: string ;
    price_CPU: number;
    quantity_CPU: number;
    number_Cores: number;
    purity_CPU: string ;
}
// ...
export interface Update_AccessoriesDto {
    iD_Accessories: number;
    сharacteristics_info: string;
}
// ...
export interface Creat_DeliveryDto {
    iD_Accessories: number;
    iD_Сomponent: string;
    name_Сomponent: string;
    type_Сomponent: string;
    price_Сomponent: number;
    quantity_Сomponent: number;
    issue_date:string;
    name_Company: string;
}